moved components into separate directories, fixed typescript types, refactored code to more meaningful and readable state, added comments
